# internal/api/modelapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Lineage Model API",
    "version": "v1",
    "description": "Machine-facing Model API for the Lineage AI model registry: publish, resolve, fetch, and manage models, versions, artifacts, lineage, and deployments. Auth is handled by infrastructure; a trusted X-Lineage-Actor header is recorded for audit."
  },
  "servers": [
    {
      "url": "/",
      "description": "Model API port (:8081)"
    }
  ],
...
  "paths": {
    "/v1/models": {
      "get": {
        "operationId": "listModels",
        "tags": [
          "models"
        ],
        "summary": "List models",
        "parameters": [
          {
            "$ref": "#/components/parameters/pageSize"
          },
          {
            "$ref": "#/components/parameters/pageToken"
          },
          {
            "$ref": "#/components/parameters/orderBy"
          },
          {
            "$ref": "#/components/parameters/q"
          },
          {
            "name": "state",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "ACTIVE",
                "ARCHIVED"
              ]
            }
          },
          {
            "name": "euSystemRiskClass",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/EUSystemRiskClass"
            },
            "description": "Stored column, pushed into SQL. Selects only models classified this way under the regime — never models nobody has classified."
          },
          {
            "name": "euGpaiTier",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/EUGpaiTier"
            }
          },
          {
            "name": "classificationState",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/ClassificationState"
            },
            "description": "Computed on every read, so this is not a stored filter. `unclassified` returns models with no row for the regime; it is not a kind of `stale`."
          },
          {
            "name": "mrmTier",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/MRMTier"
            },
            "description": "Stored column, pushed into SQL. Selects only models with an `mrm` row at this tier — never models nobody has tiered."
          },
          {
            "name": "mrmState",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/MRMState"
            },
            "description": "Computed on every read (§20.7). `untiered` returns models with no `mrm` row or an explicit `untiered`. Combined with an EU filter, returns the models that pass both."
          },
          {
            "name": "regime",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/Regime"
            },
            "description": "Turns on that regime's lens without filtering on it: `eu_ai_act` embeds `classification`, `mrm` embeds `mrm`."
          },
          {
            "name": "include",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "classification",
                "mrm"
              ]
            },
            "description": "`classification` embeds each model's EU row, `mrm` its MRM row, without filtering on either; `classification,mrm` embeds both. Omitted entirely, this endpoint runs the same query it always has — the joins and their version aggregates are not paid for unless asked for."
          }
        ],
        "responses": {
          "200": {
            "description": "Page of models",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ModelPage"
                    },
                    {
                      "$ref": "#/components/schemas/ModelInventoryPage"
                    }
                  ]
                }
              }
            }
          }
        },
        "description": "Returns ModelInventoryPage instead when any classification parameter or include=classification is present."
      },
      "post": {
        "operationId": "createModel",
        "tags": [
          "models"
        ],
        "summary": "Create a model",
        "parameters": [
          {
            "$ref": "#/components/parameters/idempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateModel"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Model"
                }
              }
            }
          },
          "409": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
    "/v1/models/{model}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/model"
        }
      ],
      "get": {
        "operationId": "getModel",
        "tags": [
          "models"
        ],
        "summary": "Get a model",
        "responses": {
          "200": {
            "description": "Model",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Model"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/Problem"
          }
        }
      },
      "patch": {
        "operationId": "patchModel",
        "tags": [
          "models"
        ],
        "summary": "Update mutable model fields",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Model"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteModel",
        "tags": [
          "models"
        ],
        "summary": "Delete a model (cascades)",
        "parameters": [
          {
            "$ref": "#/components/parameters/force"
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "409": {
            "description": "Refused. details.reason is legal_hold (with heldSince, heldBy and, when inherited, heldSubject), retention_floor (with floorDays, ageDays), or the pre-existing production-version guard. ?force=true overrides only the production-version guard — never a legal hold and never the retention floor.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/models/{model}:archive": {
      "parameters": [
        {
          "$ref": "#/components/parameters/model"
        }
      ],
      "post": {
        "operationId": "archiveModel",
        "tags": [
          "models"
        ],
        "summary": "Archive a model (soft, reversible via PATCH)",
        "responses": {
          "200": {
            "description": "Archived",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Model"
                }
              }
            }
          }
        }
      }
    },
    "/v1/models/{model}/resolve": {
      "parameters": [
        {
          "$ref": "#/components/parameters/model"
        }
      ],
      "get": {
        "operationId": "resolve",
        "tags": [
          "delivery"
        ],
        "summary": "Resolve a model to native + signed artifact refs",
        "parameters": [
          {
            "name": "stage",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "draft",
                "staging",
                "production",
                "archived"
              ]
            }
          },
          {
            "name": "version",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Resolution",
            "headers": {
              "ETag": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Resolution"
                }
              }
            }
          },
          "304": {
            "description": "Not modified (If-None-Match matched)"
          },
          "409": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
    "/v1/models/{model}/audit": {
      "parameters": [
        {
          "$ref": "#/components/parameters/model"
        }
      ],
      "get": {
        "operationId": "modelAudit",
        "tags": [
          "audit"
        ],
        "summary": "Audit trail for a model",
        "parameters": [
          {
            "$ref": "#/components/parameters/pageSize"
          },
          {
            "$ref": "#/components/parameters/pageToken"
          }
        ],
        "responses": {
          "200": {
            "description": "Page of audit events",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditPage"
                }
              }
            }
          }
        }
      }
    },
    "/v1/models/{model}/versions": {
      "parameters": [
        {
          "$ref": "#/components/parameters/model"
        }
      ],
      "get": {
        "operationId": "listVersions",
        "tags": [
          "versions"
        ],
        "summary": "List versions",
        "parameters": [
          {
            "$ref": "#/components/parameters/pageSize"
          },
          {
            "$ref": "#/components/parameters/pageToken"
          },
          {
            "name": "stage",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Page of versions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VersionPage"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "publishVersion",
        "tags": [
          "versions"
        ],
        "summary": "Publish a version (with optional inline artifacts)",
        "parameters": [
          {
            "$ref": "#/components/parameters/idempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishVersion"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishResult"
                }
              }
            }
          },
          "409": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
    "/v1/models/{model}/versions/{version}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/model"
        },
        {
          "$ref": "#/components/parameters/version"
        }
      ],
      "get": {
        "operationId": "getVersion",
        "tags": [
          "versions"
        ],
        "summary": "Get a version",
        "responses": {
          "200": {
            "description": "Version",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelVersion"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/Problem"
          }
        }
      },
      "patch": {
        "operationId": "patchVersion",
        "tags": [
          "versions"
        ],
        "summary": "Update mutable version metadata",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchVersion"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelVersion"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteVersion",
        "tags": [
          "versions"
        ],
        "summary": "Delete a version (guarded against production)",
        "parameters": [
          {
            "$ref": "#/components/parameters/force"
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "409": {
            "description": "Refused. details.reason is legal_hold (with heldSince, heldBy and, when inherited, heldSubject), retention_floor (with floorDays, ageDays), or the pre-existing production-version guard. ?force=true overrides only the production-version guard — never a legal hold and never the retention floor.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/models/{model}/versions/{version}:transition": {
      "parameters": [
        {
          "$ref": "#/components/parameters/model"
        },
        {
          "$ref": "#/components/parameters/version"
        }
      ],
      "post": {
        "operationId": "transitionVersion",
        "tags": [
          "versions"
        ],
        "summary": "Move a version to a new stage",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Transition"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Transitioned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelVersion"
                }
              }
            }
          },
          "409": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
    "/v1/models/{model}/versions/{version}/artifacts": {
      "parameters": [
        {
          "$ref": "#/components/parameters/model"
        },
        {
          "$ref": "#/components/parameters/version"
        }
      ],
      "get": {
        "operationId": "listArtifacts",
        "tags": [
          "artifacts"
        ],
        "summary": "List artifacts on a version",
        "responses": {
          "200": {
            "description": "Artifacts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Artifact"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "registerArtifact",
        "tags": [
          "artifacts"
        ],
        "summary": "Register an artifact by reference",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArtifactInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Artifact"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
...
      "pageSize": {
        "name": "pageSize",
        "in": "query",
        "schema": {
          "type": "integer",
          "default": 50,
          "maximum": 500
        }
      },
      "pageToken": {
        "name": "pageToken",
        "in": "query",
        "schema": {
          "type": "string"
        }
      },
      "orderBy": {
        "name": "orderBy",
        "in": "query",
        "schema": {
          "type": "string"
        }
      },
      "q": {
        "name": "q",
        "in": "query",
        "schema": {
          "type": "string"
        },
        "description": "Case-insensitive substring match on name"
      },
...
    },
...
    "schemas": {
...
      "ModelPage": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Model"
            }
          },
          "nextPageToken": {
            "type": "string"
          }
        }
      },
...
      "EUSystemRiskClass": {
        "type": "string",
        "enum": [
          "unclassified",
          "minimal",
          "limited",
          "high_annex_iii",
          "high_annex_i",
          "prohibited"
        ],
        "description": "The risk class of the systems this model is used in, as declared by a person. `unclassified` is a real answer meaning nobody has said yet — it is never rendered as, and never quietly becomes, `minimal`."
      },
      "EUGpaiTier": {
        "type": "string",
        "enum": [
          "none",
          "gpai",
          "gpai_systemic"
        ],
        "description": "The model itself under Art. 53/55. A general-purpose model is not 'high risk' or 'low risk' — that is a different question, answered by euSystemRiskClass."
      },
...
    }
  }
}